Add files via upload
Mea culpa. The audio.h does all the heavy lifting and should be oriented towards objects and converted into a library.  This is issue 7 of 9 or so and it is the one we used during the first robot dance contest. It is only using three of the seven bands currently but there is a version that attempts to find beats in all bands.  I think this is a good direction for the future but did not have enough time to make it work for Arduino day.